move getVersionByStore from Arc to ArcInfo
