+++ src/runtime/relevance.ts
@@ -21,7 +21,7 @@ export class Relevance {
 
   static create(arc: Arc, recipe: Recipe): Relevance {
     const relevance = new Relevance();
-    const versionByStore = arc.getVersionByStore({includeArc: true, includeContext: true});
+    const versionByStore = arc.arcInfo.getVersionByStore({includeArc: true, includeContext: true});
     recipe.handles.forEach(handle => {
       if (handle.id && versionByStore[handle.id] !== undefined) {
         relevance.versionByStore[handle.id] = versionByStore[handle.id];
